Improve comparing mechanism.
Brunch with proper logger messages instead of throwing in case of comparing with null, in case of comparing with closed document, in case of comparing while no document was opened.
Separate comparing logic and showing the results in RupsController class.
Add possibility of comparing, but not showing the results from Rups class.

DEVSIX-835