Add two methods for registering observer to RupsController
NPE is occured in highlightChanges method of PdfReaderController instance
when I invoke highlightLastSavedChanges method of Rups instance after
invoke loadDocumentFromRawContent of Rups instance.
This is because highlightLastSavedChanges method is invoked before
finished method of ObjectLoader is invoked.
So I add a listener to notice when the finish method is done.

DEVSIX-2078